add the ClassLoaderTypeSolver
