+++ javaparser-symbol-solver-core/src/main/java/com/github/javaparser/symbolsolver/resolution/typesolvers/ReflectionTypeSolver.java
@@ -16,13 +16,6 @@
 
 package com.github.javaparser.symbolsolver.resolution.typesolvers;
 
-import com.github.javaparser.resolution.declarations.ResolvedReferenceTypeDeclaration;
-import com.github.javaparser.symbolsolver.model.resolution.SymbolReference;
-import com.github.javaparser.symbolsolver.model.resolution.TypeSolver;
-import com.github.javaparser.symbolsolver.reflectionmodel.ReflectionFactory;
-
-import java.util.Optional;
-
 /**
  * Uses reflection to resolve types.
  * Classes on the classpath used to run your application will be found.
@@ -30,9 +23,7 @@ import java.util.Optional;
  *
  * @author Federico Tomassetti
  */
-public class ReflectionTypeSolver implements TypeSolver {
-
-    private TypeSolver parent;
+public class ReflectionTypeSolver extends ClassLoaderTypeSolver {
 
     private final boolean jreOnly;
 
@@ -42,6 +33,7 @@ public class ReflectionTypeSolver implements TypeSolver {
      * If false, will resolve any kind of type.
      */
     public ReflectionTypeSolver(boolean jreOnly) {
+        super(ReflectionTypeSolver.class.getClassLoader());
         this.jreOnly = jreOnly;
     }
 
@@ -54,62 +46,8 @@ public class ReflectionTypeSolver implements TypeSolver {
     }
 
     @Override
-    public TypeSolver getParent() {
-        return parent;
-    }
-
-    @Override
-    public void setParent(TypeSolver parent) {
-        this.parent = parent;
-    }
-
-    @Override
-    public SymbolReference<ResolvedReferenceTypeDeclaration> tryToSolveType(String name) {
-        if (!jreOnly || (name.startsWith("java.") || name.startsWith("javax."))) {
-            try {
-                ClassLoader classLoader = ReflectionTypeSolver.class.getClassLoader();
-
-                // Some implementations could return null when the class was loaded through the bootstrap classloader
-                // see https://docs.oracle.com/javase/8/docs/api/java/lang/Class.html#getClassLoader--
-                if (classLoader == null) {
-                    throw new RuntimeException("The ReflectionTypeSolver has been probably loaded through the bootstrap class loader. This usage is not supported by the JavaSymbolSolver");
-                }
-
-                Class<?> clazz = classLoader.loadClass(name);
-                return SymbolReference.solved(ReflectionFactory.typeDeclarationFor(clazz, getRoot()));
-            } catch (NoClassDefFoundError e) {
-                // We can safely ignore this one because it is triggered when there are package names which are almost the
-                // same as class name, with the exclusion of the case.
-                // For example:
-                // java.lang.NoClassDefFoundError: com/github/javaparser/printer/ConcreteSyntaxModel
-                // (wrong name: com/github/javaparser/printer/concretesyntaxmodel)
-                // note that this exception seems to be thrown only on certain platform (mac yes, linux no)
-                return SymbolReference.unsolved(ResolvedReferenceTypeDeclaration.class);
-            } catch (ClassNotFoundException e) {
-                // it could be an inner class
-                int lastDot = name.lastIndexOf('.');
-                if (lastDot == -1) {
-                    return SymbolReference.unsolved(ResolvedReferenceTypeDeclaration.class);
-                } else {
-                    String parentName = name.substring(0, lastDot);
-                    String childName = name.substring(lastDot + 1);
-                    SymbolReference<ResolvedReferenceTypeDeclaration> parent = tryToSolveType(parentName);
-                    if (parent.isSolved()) {
-                        Optional<ResolvedReferenceTypeDeclaration> innerClass = parent.getCorrespondingDeclaration().internalTypes()
-                                .stream().filter(it -> it.getName().equals(childName)).findFirst();
-                        if (innerClass.isPresent()) {
-                            return SymbolReference.solved(innerClass.get());
-                        } else {
-                            return SymbolReference.unsolved(ResolvedReferenceTypeDeclaration.class);
-                        }
-                    } else {
-                        return SymbolReference.unsolved(ResolvedReferenceTypeDeclaration.class);
-                    }
-                }
-            }
-        } else {
-            return SymbolReference.unsolved(ResolvedReferenceTypeDeclaration.class);
-        }
+    protected boolean filterName(String name) {
+        return !jreOnly || (name.startsWith("java.") || name.startsWith("javax."));
     }
 
 }
